scripts/requirements.txt
# Core dependencies
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
python-dotenv==1.0.1

# Database
supabase==2.3.4

# OpenAI for embeddings
openai==1.12.0

# Data processing
pandas==2.2.0

# URL parsing and robots.txt
urllib3==2.1.0

# Scheduling (optional)
schedule==1.2.1
python-crontab==3.0.0

# Link checking
linkchecker==10.3.0

# Logging
colorlog==6.8.0python-dateutil==2.8.2
